+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix issues caused by new yarn `workspaces` semantics (#71, #67).
+
 ## 1.4.0 (2018-2-16)
 
 * Add `--relative-time` to all commands, shortening the date column in logs by